Add files via upload
Problem: Having trouble recognizing $memberId and $ministryId in the same View($memberId) function. So I created a separate function called ViewMinistry($ministryId) function to handle the problem, but not it does not recognize $memberId. These are all in the Members.php Controller. The purpose of this is to print the ministry names so it can be displayed and selected for the modal "Add to Ministry".